+++ R/plot.R
@@ -290,9 +290,9 @@ setMethod("barplot", "SpatRaster",
 
 
 
-shade <- function(slope, aspect, angle=45, direction=0, normalize=FALSE, filename="", ...) {
+shade <- function(slope, aspect, angle=45, direction=0, normalize=FALSE, filename="", overwrite=FALSE, ...) {
 	
-	x <- c(slope, aspect)
+	x <- c(slope[[1]], aspect[[1]])
 
 	direction <- direction[1] * pi/180
 	zenith <- (90 - angle[1]) * pi/180
@@ -306,6 +306,6 @@ shade <- function(slope, aspect, angle=45, direction=0, normalize=FALSE, filenam
 	} else {
 		fun <- function(slp, asp) { cos(slp) * cos(zenith) + sin(slp) * sin(zenith) * cos(direction-asp) }
 	}
-	lapp(x, fun=fun, filename, ...)		
+	lapp(x, fun=fun, filename=filename, overwrite=overwrite, wopt=list(...))		
 }